Reapply "[Fix] 스크립트 분리, 공격모션 수정"
This reverts commit bc7835dc78931e48f256ce81bc876c0b569c1eee.
