+++ ZooDOS/Assets/02.Scripts/Common/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 24784198392207d4dbc0fa970b6743c4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
