+++ ZooDOS/Assets/02.Scripts/Common/Interfaces/IAttackable.cs
@@ -1,12 +1,8 @@
 ﻿// 인터페이스 모음
 public interface IAttackable
 {
+    public void Attack(Unit unit);
     public Unit FindTarget();
     public bool IsInRange(Unit unit);
-}
-
-public interface IDamagable
-{
-    public void GetDamage();
 }
