+++ ZooDOS/Assets/02.Scripts/Common/Interfaces/IDamagable.cs
@@ -1,0 +1,6 @@
+﻿public interface IDamagable
+{
+    public void GetDamage(float value);
+    public void OnDeath();
+}
+
