+++ ZooDOS/Assets/02.Scripts/Common/Interfaces/IDamagable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e0e427b982e224c4eb0345e7de1caddc
