--- ZooDOS/Assets/02.Scripts/Common/Structures.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b90cd7c0e2715d34d934d83e9d25aa3f
