+++ ZooDOS/Assets/02.Scripts/Common/Structures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1225f0e5278ca8b4a9d136bf877d9c59
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
