+++ ZooDOS/Assets/02.Scripts/Common/Structures/Hp.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 97f6f067e58f3274ba64037b15a5a8f7
