+++ ZooDOS/Assets/02.Scripts/Common/Types.cs
@@ -1,7 +1,5 @@
 ﻿// enum 모음
 
-using System.Numerics;
-
 public enum PlayerUnitType
 {
     Nothing,
