+++ ZooDOS/Assets/02.Scripts/Unit/PlayerUnit/PlayerDefenderUnit.cs
@@ -14,6 +14,7 @@ public class PlayerDefenderUnit : PlayerUnit, IAttackable
 
     protected void Update()
     {
+        // 공격 딜레이
         _leftAttackTime += Time.deltaTime;
 
         if (_leftAttackTime < _atkSpeed)
