+++ ZooDOS/Assets/02.Scripts/Unit/PlayerUnit/PlayerMedicUnit.cs
@@ -6,6 +6,7 @@ public class PlayerMedicUnit : PlayerUnit, IAttackable
 
     protected void Update()
     {
+        // 공격 딜레이
         _leftAttackTime += Time.deltaTime;
 
         if (_leftAttackTime < _atkSpeed)
