+++ ZooDOS/Assets/02.Scripts/Unit/PlayerUnit/PlayerSniperUnit.cs
@@ -6,6 +6,7 @@ public class PlayerSniperUnit : PlayerUnit, IAttackable
 
     protected void Update()
     {
+        // 공격 딜레이
         _leftAttackTime += Time.deltaTime;
 
         if (_leftAttackTime < _atkSpeed)
