+++ ZooDOS/Assets/02.Scripts/Unit/PlayerUnit/PlayerUnit.cs
@@ -15,14 +15,16 @@ public class PlayerUnit : Unit
     protected float _replaceTime;
     protected PlayerUnitType _playerUnitType;
     protected TileType _tileType;
-    protected PlayerUnitAttackRange _playerUnitAttackRange;
 
-    
+
 
     /// <summary>
     /// 배치 시 실행되는 메서드
+    /// 공격 딜레이 초기화
     /// 이번 배치에서 지정된 사거리와 위치 지정
+    /// 배치되는 타일에 참조 전달
     /// 체력 초기화
+    /// 배치 애니메이션 재생
     /// </summary>
     /// <param name="attackRange"></param>
     /// <param name="placedTile"></param>
@@ -31,6 +33,8 @@ public class PlayerUnit : Unit
         _leftAttackTime = 0;
         _attackRange = attackRange;
         _placeTile = placedTile;
+        _placeTile.PlayerUnit = this;
+
         _hp.Replace();
         transform.position += Vector3.up * Constants.FALLING_POS;
         StartCoroutine(C_FallingCoroutine());
@@ -39,6 +43,7 @@ public class PlayerUnit : Unit
 
     /// <summary>
     /// 유닛 데이터 최초 초기화 함수
+    /// 
     /// </summary>
     /// <param name="playerUnitData"> Spawner에게 전달받은 유닛 데이터 </param>
     /// <param name="enemyUnitSpawner"> Spawner에게 전달받은 enemyUnitSpawner 참조 </param>
@@ -55,11 +60,6 @@ public class PlayerUnit : Unit
         _placeCost = playerUnitData.PlaceCost;
         _replaceTime = playerUnitData.ReplaceTime;
 
-        _playerUnitAttackRange = new PlayerUnitAttackRange
-            (playerUnitData.BackwardRange * Constants.MAPTILE_LENGTH + Constants.MAPTILE_LENGTH/2,
-             playerUnitData.ForwardRange * Constants.MAPTILE_LENGTH + Constants.MAPTILE_LENGTH / 2,
-             playerUnitData.SidewardRange * Constants.MAPTILE_LENGTH + Constants.MAPTILE_LENGTH / 2
-            );
     }
 
 
@@ -75,9 +75,7 @@ public class PlayerUnit : Unit
         }
         transform.position = new Vector3(transform.position.x, targetY, transform.position.z);
     }
-
-
-
+    
     public override void OnDeath()
     {
         base.OnDeath();
