+++ ZooDOS/Assets/03.Prefabs/Test/TestObject.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 47bc1a93ba166ab4581d29f9e5f75088
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
